refactor: (#30) topic 전용 properties 생성
