+++ storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/kafka/event/post/update/UpdatePostProducer.java
@@ -7,7 +7,7 @@ import org.example.stortiessearch.infrastructure.mq.kafka.util.JsonSerializer;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
-import static org.example.stortiessearch.infrastructure.mq.kafka.KafkaProperties.UPDATE_TOPIC;
+import static org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaTopicProperties.UPDATE_TOPIC;
 
 @RequiredArgsConstructor
 @Component
