+++ storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/kafka/event/viewlog/IncreasePostViewProducer.java
@@ -7,7 +7,7 @@ import org.example.stortiessearch.infrastructure.mq.kafka.util.JsonSerializer;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
-import static org.example.stortiessearch.infrastructure.mq.kafka.KafkaProperties.INCREASE_VIEW_TOPIC;
+import static org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaTopicProperties.INCREASE_VIEW_TOPIC;
 
 @RequiredArgsConstructor
 @Component
