+++ storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/kafka/properties/KafkaTopicProperties.java
@@ -1,8 +1,8 @@
-package org.example.stortiessearch.infrastructure.mq.kafka;
+package org.example.stortiessearch.infrastructure.mq.kafka.properties;
 
 import java.util.List;
 
-public class KafkaProperties {
+public class KafkaTopicProperties {
     public static final String DELETE_TOPIC = "delete-post";
 
     public static final String UPDATE_TOPIC = "update-post";
@@ -29,12 +29,4 @@ public class KafkaProperties {
 
     public static final String RETRY_TOPIC = "retry";
 
-    public static final String GROUP_ID = "post-group";
-
-    public static final String CONTAINER_FACTORY = "kafkaListenerContainerFactoryWithManualAck";
-
-    public final static String INDEX_NAME = "storties_post_index";
-
-    public KafkaProperties() {
-    }
 }
