+++ storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/kafka/system/deadletter/DeadLetterConsumer.java
@@ -8,9 +8,9 @@ import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
 
-import static org.example.stortiessearch.infrastructure.mq.kafka.KafkaProperties.CONTAINER_FACTORY;
-import static org.example.stortiessearch.infrastructure.mq.kafka.KafkaProperties.DEAD_LETTER_TOPIC;
-import static org.example.stortiessearch.infrastructure.mq.kafka.KafkaProperties.GROUP_ID;
+import static org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaProperties.CONTAINER_FACTORY;
+import static org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaProperties.GROUP_ID;
+import static org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaTopicProperties.DEAD_LETTER_TOPIC;
 
 @Component
 @RequiredArgsConstructor
