+++ storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/kafka/system/retry/KafkaRetryProducer.java
@@ -6,7 +6,7 @@ import org.example.stortiessearch.infrastructure.mq.kafka.dto.KafkaEvent;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
-import static org.example.stortiessearch.infrastructure.mq.kafka.KafkaProperties.RETRY_TOPIC;
+import static org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaTopicProperties.RETRY_TOPIC;
 
 @Component
 @RequiredArgsConstructor
