+++ storties-search/src/main/java/org/example/stortiessearch/infrastructure/mq/kafka/system/retry/KafkaRetryScheduler.java
@@ -3,13 +3,15 @@ package org.example.stortiessearch.infrastructure.mq.kafka.system.retry;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.stortiessearch.infrastructure.mq.kafka.KafkaProperties;
+import org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaProperties;
 import org.example.stortiessearch.infrastructure.mq.kafka.dto.KafkaEvent;
 import org.example.stortiessearch.infrastructure.mq.kafka.util.JsonSerializer;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+
+import static org.example.stortiessearch.infrastructure.mq.kafka.properties.KafkaTopicProperties.RETRY_TARGET_TOPICS;
 
 @Slf4j
 @Component
@@ -24,14 +26,14 @@ public class KafkaRetryScheduler {
 
     @Scheduled(fixedDelay = 30000)
     public void processShortQueue() {
-        for (String topic : KafkaProperties.RETRY_TARGET_TOPICS) {
+        for (String topic : RETRY_TARGET_TOPICS) {
             process("redis:short:" + topic);
         }
     }
 
     @Scheduled(fixedDelay = 60000)
     public void processLongQueue() {
-        for (String topic : KafkaProperties.RETRY_TARGET_TOPICS) {
+        for (String topic : RETRY_TARGET_TOPICS) {
             process("redis:long:" + topic);
         }
     }
